Resolve warnings

+++ lib/pipeline/fragmentation/fragmenter.cpp
@@ -29,7 +29,7 @@ Packet Fragmenter::create_packet() {
             id : message.id,
             fragment_num : i,
             checksum : 0,
-            flags : last_fragment ? END : 0,
+            flags : static_cast<uint8_t>(last_fragment ? END : 0),
             type : message.type,
         },
         message_data : 0
